refactor(export): extract shared markdown/blob helpers to lib/export-utils.ts
