+++ apps/web-ui/lib/skill-export.ts
@@ -9,21 +9,7 @@
  */
 
 import type { SkillDTO } from "@/lib/client-skill-service";
-
-/**
- * Wrap content in a code fence that cannot be terminated early: the fence is
- * one backtick longer than the longest backtick run inside the content.
- */
-function fence(content: string, lang = "markdown"): string {
-    const longestRun = content.match(/`+/g)?.reduce((max, run) => Math.max(max, run.length), 0) ?? 0;
-    const marker = "`".repeat(Math.max(3, longestRun + 1));
-    return `${marker}${lang}\n${content}\n${marker}`;
-}
-
-/** GitHub-style heading anchor: lowercase, trim, spaces → hyphens, drop non-alnum. */
-function anchor(text: string): string {
-    return text.toLowerCase().trim().replace(/[^\w\s-]/g, "").replace(/\s+/g, "-");
-}
+import { fence, anchor, downloadBlob, downloadText, fileSafe, yamlScalar } from "@/lib/export-utils";
 
 /** Build the Markdown document for a single skill. Pure. */
 export function buildSkillMarkdown(skill: SkillDTO): string {
@@ -71,41 +57,6 @@ export function buildAllSkillsMarkdown(skills: SkillDTO[]): string {
     header.push("", "---", "");
     const body = sorted.map((s) => `${buildSkillMarkdown(s)}\n---\n`);
     return `${header.join("\n")}\n${body.join("\n")}`;
-}
-
-/** Trigger a client-side download of a Blob. */
-function downloadBlob(blob: Blob, filename: string): void {
-    const url = URL.createObjectURL(blob);
-    const link = document.createElement("a");
-    link.href = url;
-    link.download = filename;
-    document.body.appendChild(link);
-    link.click();
-    document.body.removeChild(link);
-    URL.revokeObjectURL(url);
-}
-
-/** Trigger a client-side download of `content` as `filename`. */
-function downloadText(content: string, filename: string, mimeType = "text/markdown;charset=utf-8"): void {
-    downloadBlob(new Blob([content], { type: mimeType }), filename);
-}
-
-function fileSafe(text: string): string {
-    return text.toLowerCase().replace(/[^a-z0-9]+/g, "-").replace(/^-+|-+$/g, "") || "skill";
-}
-
-/**
- * Emit a YAML scalar that is safe to embed in frontmatter. Single-line values use
- * double-quoted form (backslash + double-quote escaped); multi-line values use a
- * block scalar so newlines are preserved verbatim.
- */
-function yamlScalar(value: string): string {
-    const v = value ?? "";
-    if (v.includes("\n")) {
-        const indented = v.split("\n").map((l) => `  ${l}`).join("\n");
-        return `|-\n${indented}`;
-    }
-    return `"${v.replace(/\\/g, "\\\\").replace(/"/g, '\\"')}"`;
 }
 
 /**
